마이페이지 오류 수정
