--- src/components/page/myPage/index.js
@@ -1,2 +1,0 @@
-import User from "./User";
-export { User };
